Add results wrapper component
This component is in place to add a loading spinner
in an accessible way. For this to happen, we need
it wrapping the results container so that its
div with aria-loading attribute set is present
all the time (with either a "true" or false" value).
Otherwise it won't be useful for screen readers.

https://dockyard.com/blog/2020/03/02/accessible-loading-indicatorswith-no-extra-elements